backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
pydantic==1.10.7
python-dotenv==1.0.0
requests>=2.28.0
cryptography>=3.4.0
